Modularized the widgets for the user profile screen
